% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{list_dataset}
\alias{list_dataset}
\title{List the Data in the Dataset}
\usage{
list_dataset(id, src, ...)
}
\arguments{
\item{id}{string. The id of the dataset.}

\item{src}{string. The name of the source platform. Possible values are one of:
\code{c("Taiwan", "Taipei")} or the abbreviation: \code{c("tw", "tp")}.}

\item{...}{Additional arguments which will be passed to the specified source platform.}
}
\description{
List the Data in the Dataset
}
\examples{
\dontrun{
 list_dataset("8ef1626a-892a-4218-8344-f7ac46e1aa48", "Taipei")
}
}

