% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{search_dataset}
\alias{search_dataset}
\title{Search a Dataset}
\usage{
search_dataset(keyword, src = "Taiwan", ...)
}
\arguments{
\item{keyword}{string. The keyword to search.}

\item{src}{string. The name of the source platform. Possible values are one of:
\code{c("Taiwan", "Taipei")} or the abbreviation: \code{c("tw", "tp")}.}

\item{...}{Additional arguments which will be passed to the specified source platform.}
}
\description{
Search a Dataset
}

