% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{download_data}
\alias{download_data}
\title{Download the Data from the Source Platform.}
\usage{
download_data(rid, src, format = "table", ...)
}
\arguments{
\item{rid}{string. The id of the data.}

\item{src}{string. The name of the source platform. Possible values are one of:
\code{c("Taiwan", "Taipei")} or the abbreviation: \code{c("tw", "tp")}.}

\item{...}{Additional arguments which will be passed to the specified source platform.}
}
\description{
Download the Data from the Source Platform.
}

